requirements.txt
google-cloud-vision
django